+++ adapter-module/src/main/kotlin/com/reservation/persistence/restaurant/repository/dsl/FindRestaurantRepository.kt
@@ -4,12 +4,12 @@ import com.querydsl.core.group.GroupBy.groupBy
 import com.querydsl.core.group.GroupBy.list
 import com.querydsl.core.types.Projections
 import com.querydsl.jpa.impl.JPAQueryFactory
-import com.reservation.persistence.restaurant.QRestaurantCuisinesEntity.restaurantCuisinesEntity
-import com.reservation.persistence.restaurant.QRestaurantEntity.restaurantEntity
-import com.reservation.persistence.restaurant.QRestaurantNationalitiesEntity.restaurantNationalitiesEntity
-import com.reservation.persistence.restaurant.QRestaurantPhotoEntity.restaurantPhotoEntity
-import com.reservation.persistence.restaurant.QRestaurantTagsEntity.restaurantTagsEntity
-import com.reservation.persistence.restaurant.QRestaurantWorkingDayEntity.restaurantWorkingDayEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantCuisinesEntity.restaurantCuisinesEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantEntity.restaurantEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantNationalitiesEntity.restaurantNationalitiesEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantPhotoEntity.restaurantPhotoEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantTagsEntity.restaurantTagsEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantWorkingDayEntity.restaurantWorkingDayEntity
 import com.reservation.restaurant.port.output.FindRestaurant
 import com.reservation.restaurant.port.output.FindRestaurant.FindRestaurantPhoto
 import com.reservation.restaurant.port.output.FindRestaurant.FindRestaurantResult
