+++ adapter-module/src/main/kotlin/com/reservation/persistence/restaurant/repository/dsl/RestaurantPhotoQuerySpec.kt
@@ -1,7 +1,7 @@
 package com.reservation.persistence.restaurant.repository.dsl
 
 import com.querydsl.core.types.dsl.BooleanExpression
-import com.reservation.persistence.restaurant.QRestaurantPhotoEntity.restaurantPhotoEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantPhotoEntity.restaurantPhotoEntity
 
 object RestaurantPhotoQuerySpec {
     fun photoRestaurantIdentifiersIn(identifiers: Set<String>): BooleanExpression =
