+++ adapter-module/src/main/kotlin/com/reservation/persistence/restaurant/repository/dsl/RestaurantWorkingDaysQuerySpec.kt
@@ -1,7 +1,7 @@
 package com.reservation.persistence.restaurant.repository.dsl
 
 import com.querydsl.core.types.dsl.BooleanExpression
-import com.reservation.persistence.restaurant.QRestaurantWorkingDayEntity.restaurantWorkingDayEntity
+import com.reservation.persistence.restaurant.entity.QRestaurantWorkingDayEntity.restaurantWorkingDayEntity
 
 object RestaurantWorkingDaysQuerySpec {
     fun workingDayRestaurantIdentifiersIn(identifiers: Set<String>): BooleanExpression =
